+++ lib/ui/menu/menu_screen.dart
@@ -28,7 +28,7 @@ class MenuScreen extends ConsumerWidget {
           'NESd',
           style: TextStyle(
             color: Theme.of(context).colorScheme.primary,
-            fontWeight: FontWeight.bold,
+            fontVariations: const [FontVariation.weight(700)],
           ),
         ),
       ),