+++ lib/ui/settings/settings_screen.dart
@@ -61,7 +61,7 @@ class SettingsScreen extends HookConsumerWidget {
           'Settings',
           style: TextStyle(
             color: Theme.of(context).colorScheme.primary,
-            fontWeight: FontWeight.bold,
+            fontVariations: const [FontVariation.weight(700)],
           ),
         ),
       ),