+++ test/ui/robot.dart
@@ -75,6 +75,15 @@ class Robot extends BaseRobot {
       buildNumber: '1337',
     );
 
+    await _loadFont('Inter', ['assets/fonts/Inter-Regular.ttf']);
+    await _loadFont('Ubuntu Mono', [
+      'assets/fonts/UbuntuMono-Regular.ttf',
+      'assets/fonts/UbuntuMono-Italic.ttf',
+    ]);
+    await _loadFont('MaterialIcons', [
+      '${Platform.environment['FLUTTER_ROOT']}/bin/cache/artifacts/material_fonts/MaterialIcons-Regular.otf',
+    ]);
+
     tester.view.physicalSize =
         const Size(1920, 1080) * tester.view.devicePixelRatio;
 
@@ -106,5 +115,17 @@ class Robot extends BaseRobot {
         File(filename).writeAsBytesSync(bytes.buffer.asUint8List());
       }
     });
+  }
+
+  Future<void> _loadFont(String family, List<String> fontFiles) async {
+    final fontLoader = FontLoader(family);
+
+    for (final fontFile in fontFiles) {
+      final fontData = rootBundle.load(fontFile);
+
+      fontLoader.addFont(fontData);
+    }
+
+    await fontLoader.load();
   }
 }