+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added licenses for the used fonts
+
 ### Changed
 
+- Switched from the default font to Inter
 - Improved test coverage so issues can be found more quickly
 
 ## [0.11.0] - 2025-04-06